change ordering for parameters, lines for checker

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 
-from typing import List, Union
-
 """Defines sum_mixed_list function"""
 
+from typing import List, Union
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """Adds together all values (floats or integers) in a list
 
     Args:
